fourth commit

+++ src/main/java/com/paves/Repository/PermissionRepository.java
@@ -1,9 +1,16 @@
 package com.paves.Repository;
 
 import com.paves.Entity.Permission;
+import com.paves.Entity.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface PermissionRepository extends JpaRepository<Permission, Integer> {
+public interface PermissionRepository extends JpaRepository<Permission, Long> {
+
+
 }
